requirements.txt
openai>=1.30.0
pydantic>=2.0.0

# PDF parsing — layout detection model (detectron2) runs under hi_res strategy
unstructured[pdf]>=0.13.0
unstructured-inference>=0.7.0
pikepdf>=8.0.0
Pillow>=10.0.0

# Vector store
faiss-cpu>=1.8.0
numpy>=1.26.0

# BM25 keyword retrieval
rank-bm25>=0.2.2

python-dotenv>=1.0.0
